Bugfix: read ren. scale only from HMIX block
A SLHA output file might contain output blocks with DR-bar parameters
at multiple scales.  SPheno for example is a spectrum generator which
prints the gauge and Yukawa couplings at MX and MSUSY.

The block, which GM2Calc assumes to be there is HMIX, because it
specifies tan(beta).  tan(beta) is a low-energy parameter and SPheno
does not print the content of HMIX at MX.  Therefore, the HMIX block
is most probably the block which is output at the SUSY scale.

Note: If there are multiple HMIX blocks, we need to reconsider our
SLHA reading procedure.  We'll need to decide at which ren. scale we
want the parameters to be defined.