adding function estimating the uncertainty for amu(2-loop)
using Eq. (84) of arXiv:hep-ph/0609168 .

Note:

Eq (84), arXiv:hep-ph/0609168 takes into account photonic 2-loop
corrections and 2-loop fermion/sfermion corrections, which were
unknown at the time arXiv:hep-ph/0609168 was published.  However,
GM2Calc includes the complete photonic 2-loop corrections and
approximations for the 2-loop fermion/sfermion corrections.  For this
reason it is debatable whether Eq. (84) (especially the finite offset)
is applicable for the 2-loop calculation of GM2Calc.

For this reason, the function is not used so far.