stop DR-bar -> on-shell conversion iteration in the MSSM earlier
as soon as precision of convergence becomes worse.

There are points where the iteration does not converge (and has not
been stopped so far) due to periodic cycles, where some parameters
(mu, M1, M2) simply exchange their position.  In such scenarios the
precision of the convergence does not increase over the cycles.

With this commit we stop the iteration as soon as we observe that the
precision of convergence becomes worse.  This can lead to a massive
performance improvement when randomly scanning over the MSSM parameter
space.